Fixed virtctl console / vnc
Get current namespace dynamically
