+++ pkg/virt-operator/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "//pkg/certificates:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/service:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/virt-controller/leaderelectionconfig:go_default_library",
         "//pkg/virt-operator/creation/components:go_default_library",
         "//pkg/virt-operator/install-strategy:go_default_library",
@@ -20,6 +19,7 @@ go_library(
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
+        "//staging/src/kubevirt.io/client-go/util:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus/promhttp:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
         "//vendor/k8s.io/api/apps/v1:go_default_library",
