+++ staging/src/kubevirt.io/client-go/kubecli/BUILD.bazel
@@ -53,6 +53,7 @@ go_library(
     deps = [
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/subresources:go_default_library",
+        "//staging/src/kubevirt.io/client-go/util:go_default_library",
         "//staging/src/kubevirt.io/client-go/version:go_default_library",
         "//vendor/github.com/golang/mock/gomock:go_default_library",
         "//vendor/github.com/gorilla/websocket:go_default_library",
